+++ src/main/kotlin/com/calmwolfs/bedwar/config/features/GuiConfig.java
@@ -6,7 +6,7 @@ import io.github.moulberry.moulconfig.annotations.*;
 import org.lwjgl.input.Keyboard;
 
 public class GuiConfig {
-    @ConfigOption(name = "Edit GUI Locations", desc = "Change the position of the mods overlays")
+    @ConfigOption(name = "Edit GUI Locations", desc = "Change the position of the mods overlays. §eCommand:/bw gui")
     @ConfigEditorButton(buttonText = "Edit")
     public Runnable positions = GuiEditorManager::openGuiPositionEditor;
 
@@ -16,7 +16,7 @@ public class GuiConfig {
     public int keyBindOpen = Keyboard.KEY_NONE;
 
     @Expose
-    @ConfigOption(name = "Global GUI scale", desc = "Globally scale all mods GUIs")
+    @ConfigOption(name = "Global GUI scale", desc = "Globally scale all of the mods GUIs")
     @ConfigEditorSlider(minValue = 0.1F, maxValue = 10, minStep = 0.05F)
     public float globalScale = 1F;